\name{alizadeh}
\alias{alizadeh}
\docType{data}
\title{Loads sample data from a lymphoma/leukemia gene expression study}
\description{The sample data for the MCFS-ID algorithm}
\usage{data(alizadeh)}

\format{
A 4026 x 62 gene expression data matrix of log-ratio values. The last column
contains the annotations of the 62 samples with respect to the cancer types
C, D, F.
}
\source{
The data are from the lymphoma/leukemia study of A. Alizadeh et al., Nature 403:503-511 (2000), http://llmpp.nih.gov/lymphoma/index.shtml
}
\keyword{datasets}
